Check the invariant in mkTopTCvSubst
Summary:
This came up when I tried to fix `coreView` in `Type`.
`mkTopTCvSubst` is called where the range of the substitution has free
variables. I added an assert and replaced the violators with
`mkOpenTCvSubstPrs` which has the same type signature, but also adds the
type variables of the range of the substitution to the in-scope set.

Test Plan: ./validate

Reviewers: simonpj, goldfire, bgamari, austin

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D1801

GHC Trac Issues: #11371