Make quantifyTyVars and kindGeneralize deterministic
Summary:
quantifyTyVars and kindGeneralize used to take non-deterministic sets of
TyCoVars and then convert them to lists to be wrapped in a forall.
This manifested itself as non-deterministic inferred type signatures and
impacted optimizations like worker-wrapper.
This does all the plumbing, so that the TyCoVars are passed in
deterministic sets.

Test Plan:
I've added a test that reproduces the issue
./validate

Reviewers: simonpj, austin, goldfire, bgamari

Subscribers: goldfire, thomie, simonmar

Differential Revision: https://phabricator.haskell.org/D1924

GHC Trac Issues: #4012